Delete raw encoding and ODBCCHAR.
The "raw" encoding was Python 2.7 only.

I originally created ODBCCHAR to replace SQLWCHAR because unixODBC would define it as wchar_t
even when that was 4 bytes.  The data in the buffer of 4-byte wchar_t's was still 2-byte data.
Now I've just simplified to uint16_t.  I added this to HACKING.md.

Deleted Tuple wrapper.  Use the Object wrapper and the PyTuple_ functions.  This is to prepare
for possibly using the ABI which would not allow me access the internal item pointers directly,
so I could not use operator[] to set items.  (Python has __getitem__ and __setitem__, but to
overload __setitem__ in C++ you can only return a reference to the internal data.)